refactor: move `alt` and `api` to separate functions
