+++ src/render/markdown/container/alt-container.ts
@@ -1,0 +1,23 @@
+import { findDocument } from "../../../utils/find-document";
+import { type ContainerCallback } from "./container-callback";
+import { type ContainerContext } from "./container-context";
+
+/**
+ * Create a container renderer "alt" to render alternative texts.
+ *
+ * @internal
+ */
+export function altContainer(context: ContainerContext): ContainerCallback {
+    const { md, env, docs } = context;
+    return (tokens, index) => {
+        const token = tokens[index];
+        const needle = token.info.split(" ")[1];
+        const doc = findDocument(docs, needle);
+
+        if (doc && doc.format === "markdown") {
+            return md.render(doc.body, env);
+        }
+
+        return md.render(token.content, env);
+    };
+}
