+++ src/render/markdown/container/container-callback.ts
@@ -1,0 +1,13 @@
+import type MarkdownIt from "markdown-it";
+import { type MarkdownEnv } from "../../markdown-env";
+
+/**
+ * @internal
+ */
+export type ContainerCallback = (
+    tokens: MarkdownIt.Token[],
+    index: number,
+    options: MarkdownIt.Options,
+    env: MarkdownEnv,
+    self: MarkdownIt.Renderer,
+) => string;
